Move model loading to processing core
